Refactoring custom ColumnFilter
