+++ src/components/column-filter/ColumnFilter.js
@@ -1,0 +1,44 @@
+import React from 'react';
+import {
+  FormControl,
+  InputLabel,
+  Select,
+  MenuItem,
+  Input,
+} from '@material-ui/core';
+
+function ColumnFilter({
+  filterList,
+  onChange,
+  filterIndex,
+  column,
+  optionValues,
+}) {
+  const handleChange = (event) => {
+    const value = event.target.value !== 'All' ? event.target.value : undefined;
+    if (value) onChange(value, filterIndex, column.name);
+  };
+
+  return (
+    <FormControl key={filterIndex} fullWidth>
+      <InputLabel htmlFor={column.name}>{column.label}</InputLabel>
+      <Select
+        fullWidth
+        value={filterList[filterIndex].length ? filterList[filterIndex].toString() : 'All'}
+        name={column.name}
+        onChange={handleChange}
+        input={<Input name={column.name} id={column.name} />}>
+        <MenuItem value={'All'} key={0}>
+          {'All'}
+        </MenuItem>
+        {optionValues.map((filterValue, _filterIndex) => (
+          <MenuItem value={filterValue} key={_filterIndex + 1}>
+            {filterValue != null ? filterValue.toString() : ''}
+          </MenuItem>
+        ))}
+      </Select>
+    </FormControl>
+  );
+};
+
+export default ColumnFilter;
