+++ src/components/column-filter/index.js
@@ -1,0 +1,2 @@
+import ColumnFilter from './ColumnFilter';
+export default ColumnFilter;
