+++ src/index.js
@@ -1,8 +1,10 @@
 import Cell from './components/cell';
 import Toolbar from './components/toolbar';
 import DataTable from './components/datatable';
+import ColumnFilter from './components/column-filter';
 export {
   Cell,
   Toolbar,
   DataTable,
+  ColumnFilter,
 };
